Add boolean show_already_visited_parents
True means that you can expand nodes and go around in a cycle. This
happens when you go to a child and expand its parents to find the parent
again. Or when you expand a parent and then find the child node again.

A parent X has a child Y which have a parent X which have a parent Y etc